requirements.txt
requests==2.31.0
selenium==4.18.1
Jinja2==3.1.3
PyYAML==6.0.1
APScheduler==3.10.4
python-dotenv==1.0.1
webdriver-manager==4.0.1
msal==1.28.0
